feat(worker): all metrics are computed

+++ worker_v2_impl_go/compute/types.go
@@ -1,6 +1,6 @@
 package compute
 
 type resolution struct {
-	width  int64
-	height int64
+	width  float64
+	height float64
 }
